Updating dialogflow unit tests
